// frontend/src/types/receipt.ts
export interface ReceiptData {
  logo: string | null;
  logoScale: number;
  currency: string;
  currencySymbol: string;
  receiptNumber: string;
  receiptDate: string;
  receivedFrom: string;
  amount: number;
  paymentMethod: string;
  receivedBy: string;
  reason: string;
  companyName: string;
  companyAddress: string;
  companyPhone: string;
  titleColor: string;
  amountColor: string;
}